\name{RcmdrPlugin.ROC-internal}
\title{Internal RcmdrPlugin.ROC objects}
\alias{fncvROCR}
\alias{fncvpROC}
\alias{fncPairedpROC}
\alias{fncUnpairedpROC}
\alias{fncLRpROC}
\alias{fndHosmerLemeshow}
\alias{fncReclassification}
\description{Internal RcmdrPlugin.ROC objects.}
\details{These are not to be called by the user.}
\keyword{internal}
