\name{RcmdrPlugin.ROC-package}
\alias{RcmdrPlugin.ROC-package}
\alias{RcmdrPlugin.ROC}
\docType{package}
\title{
Receiver Operator Characteristic Rcmdr Plug-In
}
\description{
Rcmdr GUI extension plug-in for Receiver Operator Characteristic tools from pROC package. Also it ads a Rcmdr GUI extension for Hosmer and Lemeshow GOF test from the package ResourceSelection.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.ROC\cr
Type: \tab Package\cr
Version: \tab 1.0-19\cr
Date: \tab 2022-05-15\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Daniel-Corneliu Leucuta, Mihaela Hedesiu, Andrei Achimas, Oana Almasan

Maintainer: Daniel-Corneliu Leucuta <danny.ldc@gmail.com>
}

\examples{
data(aSAH)
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}
}
