\name{RcmdrPlugin.ROC-package}
\alias{RcmdrPlugin.ROC-package}
\alias{RcmdrPlugin.ROC}
\docType{package}
\title{
Receiver Operator Characteristic Rcmdr Plug-In
}
\description{
This package provides a Rcmdr GUI extension plug-in for Receiver Operator Characteristic tools from pROC and ROCR packages. Also it ads a Rcmdr GUI extension for Hosmer and Lemeshow GOF test from the package ResourceSelection, reclassification table, categorical and continuous net reclassification improvement (NRI) and integrated discrimination improvement (IDI) from the PredictABEL package.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.ROC\cr
Type: \tab Package\cr
Version: \tab 1.0-16\cr
Date: \tab 2014-07-07\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Daniel-Corneliu Leucuta, Mihaela Hedesiu, Andrei Achimas, Oana Almasan

Maintainer: Daniel-Corneliu Leucuta <danny.ldc@gmail.com>
}

\examples{
data(aSAH)
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}
}